requirements.txt
# Web Framework
flask==2.3.3
streamlit==1.29.0
flask-cors==4.0.0

# Authentication
flask-jwt-extended==4.5.3
werkzeug==2.3.7

# Database
psycopg2-binary==2.9.7
pandas==2.1.4

# AI/ML Libraries
groq==0.4.1
sentence-transformers==2.7.0
transformers>=4.34.0
torch>=2.0.1
huggingface_hub>=0.16.4
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3

# File Processing
openpyxl==3.1.2
